fix b64 decoding
