+++ app/server/server.go
@@ -1,8 +1,6 @@
 package server
 
 import (
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"log"
@@ -94,22 +92,4 @@ func sendError(err error, w http.ResponseWriter) {
 	log.Println(err)
 	w.WriteHeader(http.StatusInternalServerError)
 	fmt.Fprintf(w, "%s", err)
-}
-
-// parseB64Map accepts a base64-encoded json object and
-// unmarshals it to a map of strings
-func parseB64Map(in string) (map[string]string, error) {
-	// decode from base64
-	decoded, err := base64.RawStdEncoding.DecodeString(in)
-	if err != nil {
-		return nil, err
-	}
-	out := map[string]string{}
-	// unmarshal json
-	if len(decoded) > 0 {
-		if err := json.Unmarshal(decoded, &out); err != nil {
-			return nil, err
-		}
-	}
-	return out, nil
 }
